New version with some fixes
Added support for systems without certificate.
Corrected debug message that crashed the system.
Merged the scandisk and checksub threads in one thread.
Changed the start time of the interval counter to the end of the last
search (instead of the begin).
Added some release groups.
Added exception handler for the show page of addic7ed to prevent
crashes.
Corrected a response from the chardet library.